refactoring: moves plugin code into proper namespace
