+++ lib/redmine_cc_addresses.rb
@@ -1,0 +1,8 @@
+require 'redmine_cc_addresses/hooks'
+
+module RedmineCcAddresses
+  def self.setup
+    RedmineCcAddresses::IssuePatch.apply
+    RedmineCcAddresses::MailerPatch.apply
+  end
+end
